# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: DataShare API
  description: |
    Secure file transfer platform API for freelancers and small businesses.
    Enables file upload (authenticated or anonymous), temporary download link
    generation, tag/password management and file history.
  version: 1.0.0
  contact:
    name: DataShare Team
servers:
  - url: https://localhost/api
    description: Local Docker Compose (via Nginx reverse proxy)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15min TTL by default, see JWT_EXPIRES_IN)
  parameters:
    FileId:
      in: path
      name: fileId
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Error:
      type: object
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          description: Single message, or an array of messages for class-validator errors
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: File not found
        error:
          type: string
          example: Not Found

    # ── Auth ──
...
    CreateLinkRequest:
      type: object
      properties:
        ttlSeconds:
          type: integer
          minimum: 60
          description: Defaults to DOWNLOAD_LINK_TTL_SECONDS (86400s / 24h).
          example: 86400
        maxDownloads:
          type: integer
          minimum: 0
          description: 0 or omitted = unlimited downloads.
          example: 5
    DownloadTokenResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileId:
          type: string
          format: uuid
        token:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        downloadCount:
          type: integer
        maxDownloads:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Public download ──
    DownloadInfoResponse:
      type: object
      properties:
        originalName:
          type: string
          example: report-q4.pdf
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: string
          example: "2048576"
        hasPassword:
          type: boolean

# ──────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────
paths:
...
  /files/{fileId}/links:
    post:
      tags: [Download Links]
      summary: Create a shareable download link for a file
      operationId: createDownloadLink
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Download link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadTokenResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: [Download Links]
      summary: List active (non-expired) download links for a file
      operationId: listDownloadLinks
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Active download links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadTokenResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{fileId}/links/{tokenId}:
    delete:
      tags: [Download Links]
      summary: Revoke a download link
      description: Sets the token's expiresAt to now, making it immediately unusable.
      operationId: revokeDownloadLink
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
        - in: path
          name: tokenId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Link revoked
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the file owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File or token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ════════════════════════════════════════════
  # DOWNLOAD (PUBLIC)
  # ════════════════════════════════════════════
  /download/{token}/info:
    get:
      tags: [Download]
      summary: Get download link info (US02)
      description: |
        Returns file metadata before download. Public endpoint — no authentication required.
        Shows original filename, MIME type, size, and whether a password is required.
      operationId: getDownloadInfo
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            format: uuid
          description: Unique download token (UUID v4)
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadInfoResponse"
        "404":
          description: Token not found, or the underlying file was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Link expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /download/{token}:
    get:
      tags: [Download]
      summary: Download a file (US02)
      description: |
        Streams the file directly from the backend (proxied from MinIO — not a
        redirect). Public endpoint. If the file is password-protected (US09),
        the password must be provided as a query parameter.
      operationId: downloadFile
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            format: uuid
          description: Unique download token (UUID v4)
        - in: query
          name: password
          required: false
          schema:
            type: string
          description: File password (required if the file is password-protected)
      responses:
        "200":
          description: >-
            File content, streamed with Content-Disposition attachment headers.
            Content-Type reflects the file's real stored MIME type when known,
            falling back to application/octet-stream otherwise — not always
            application/octet-stream literally.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Password required, or incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Token not found, or the underlying file was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Link expired, or download limit (maxDownloads) reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"